requirements.txt
# Google Cloud Services
google-cloud-aiplatform==1.73.0
google-cloud-firestore==2.14.0
google-cloud-logging==3.9.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.18.4

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Web Framework (for future API)
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Utilities
tenacity==8.2.3  # Retry logic
python-dateutil==2.8.2

# Phase 2 - Caching and Performance
redis==5.0.1
numpy==1.26.3
scikit-learn==1.4.0  # For clustering in vector search

# Development
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0

# Phase 2 - Caching and Performance
redis==5.0.1
numpy==1.26.3
scikit-learn==1.4.0
